backend/requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.2
sqlalchemy==1.4.46